sensor/requirements.txt
bme680==2.0.0
python-socketio==5.13.0
python-socketio[client]==5.13.0
requests==2.32.3
python-engineio~=4.12.0
dotenv==0.9.9
